fix(native): make the binding load lazy so CLI dispatch survives it
`dist/index.js` is both the `bin` entry and the package `main`, and it
re-exports the library surface (`export * from './lib.js'` -> `initDatabase`
-> `better-sqlite3-guard.js`). The guard loaded the addon at MODULE
EVALUATION, so an unloadable binding threw before `main()` could dispatch:
`--help`, `doctor`, `repair` and the MCP startup self-heal all died with exit
1 — every command that exists to fix a broken binding.

Verified against the built artefact: at base 7d8f7e8 and at a10cc0f,
`node dist/index.js --help` with a throwing better-sqlite3 exits 1 from
better-sqlite3-guard.js before dispatch. The eager chain is inherited, not
introduced by a10cc0f; a10cc0f added further static paths to the same
side-effectful module (scan-exit.ts, native-binding.ts).

Two changes:

1. Extract the pure API — `NativeModuleLoadError`, `formatNativeLoadError`,
   `isNativeModuleLoadError`, `isPackagedPrebuildLoadError` and their
   patterns — into the side-effect-free `database/native-load-classify.ts`.
   The guard re-exports it, so existing callers are unaffected; `scan-exit.ts`
   (statically imported by `src/index.ts`) now imports the pure module and has
   no path to the loader at all.

2. The guard resolves better-sqlite3 lazily and memoised via
   `getBetterSqlite3()`. The typed `NativeModuleLoadError` contract is
   unchanged — it now surfaces at first `new Database(...)`, which is where
   better-sqlite3 13 resolves its binding anyway — but importing the module
   costs nothing. `init.ts` and `migrate.ts` construct through the accessor.
