+++ src/database/init.ts
@@ -3,7 +3,8 @@
  */
 
 import type Database from 'better-sqlite3';
-import BetterSqlite3, { isNativeModuleLoadError, formatNativeLoadError } from './better-sqlite3-guard.js';
+import { getBetterSqlite3 } from './better-sqlite3-guard.js';
+import { isNativeModuleLoadError, formatNativeLoadError } from './native-load-classify.js';
 import { existsSync, mkdirSync, readFileSync, statSync, writeFileSync, unlinkSync, renameSync, copyFileSync, readdirSync, openSync, closeSync, realpathSync } from 'fs';
 import { basename, dirname, join } from 'path';
 import { homedir } from 'os';
@@ -139,7 +140,7 @@ function enforceSafeRuntimePath(expandedPath: string, explicitDbPath: boolean): 
 function inspectDatabaseFile(dbPath: string): DatabaseInspection {
   let inspectionDb: Database.Database | null = null;
   try {
-    inspectionDb = new BetterSqlite3(dbPath, {
+    inspectionDb = new (getBetterSqlite3())(dbPath, {
       readonly: true,
       fileMustExist: true,
     });
@@ -332,7 +333,7 @@ function attemptDumpRecovery(dbPath: string): Database.Database | null {
     console.error(`[database] Backed up corrupt database to: ${backupPath}`);
 
     // Create fresh database and import the dump
-    const freshDb = new BetterSqlite3(dbPath);
+    const freshDb = new (getBetterSqlite3())(dbPath);
     try {
       freshDb.exec(dumpOutput);
       console.error('[database] Successfully recovered data via dump/reimport.');
@@ -557,7 +558,7 @@ export function initDatabase(dbPath?: string): Database.Database {
   // Wrap the initial open in try/catch to handle corrupt files gracefully
   let database: Database.Database;
   try {
-    database = new BetterSqlite3(expandedPath);
+    database = new (getBetterSqlite3())(expandedPath);
   } catch (openError) {
     // A NATIVE-MODULE load failure (missing / ABI-mismatched better-sqlite3
     // binding) throws here too — better-sqlite3 resolves its binding lazily in
@@ -581,12 +582,12 @@ export function initDatabase(dbPath?: string): Database.Database {
     if (latestHealthyBackup) {
       console.error(`[database] Restoring latest healthy backup with ${latestHealthyBackup.count} memories: ${latestHealthyBackup.path}`);
       restoreBackupAsLive(expandedPath, latestHealthyBackup.path, 'failed-open');
-      database = new BetterSqlite3(expandedPath);
+      database = new (getBetterSqlite3())(expandedPath);
     } else {
       const backupPath = backupCorruptDatabase(expandedPath);
       console.error(`   Backed up to ${backupPath}`);
       console.error('   Creating fresh database...');
-      database = new BetterSqlite3(expandedPath);
+      database = new (getBetterSqlite3())(expandedPath);
     }
   }
 
@@ -611,7 +612,7 @@ export function initDatabase(dbPath?: string): Database.Database {
         if (freshIntegrityResult === 'ok') {
           console.warn('[database] Integrity failure was transient. Reopening the on-disk database without destructive recovery.');
           database.close();
-          database = new BetterSqlite3(expandedPath);
+          database = new (getBetterSqlite3())(expandedPath);
           liveIntegrityOk = true; // On-disk re-check was clean.
         } else {
           console.warn(`[database] Fresh integrity check also failed: ${freshIntegrityResult}`);
@@ -628,7 +629,7 @@ export function initDatabase(dbPath?: string): Database.Database {
             if (latestHealthyBackup) {
               console.error(`[database] Recovery failed. Restoring latest healthy backup with ${latestHealthyBackup.count} memories: ${latestHealthyBackup.path}`);
               restoreBackupAsLive(expandedPath, latestHealthyBackup.path, 'recovery-failed');
-              database = new BetterSqlite3(expandedPath);
+              database = new (getBetterSqlite3())(expandedPath);
             } else {
               // Recovery failed — backup and create fresh
               if (existsSync(expandedPath)) {
@@ -636,7 +637,7 @@ export function initDatabase(dbPath?: string): Database.Database {
                 console.error(`[database] Recovery failed. Backed up corrupt file to: ${backupPath}`);
               }
               console.error('[database] Creating fresh database...');
-              database = new BetterSqlite3(expandedPath);
+              database = new (getBetterSqlite3())(expandedPath);
             }
           }
           // Dump-recovery, backup-restore and fresh-create all yield a
@@ -673,7 +674,7 @@ export function initDatabase(dbPath?: string): Database.Database {
       console.error(`[database] Empty live database detected alongside a recent healthy backup (${latestHealthyBackup.count} memories). Restoring ${latestHealthyBackup.path}`);
       database.close();
       restoreBackupAsLive(expandedPath, latestHealthyBackup.path, 'empty-live');
-      database = new BetterSqlite3(expandedPath);
+      database = new (getBetterSqlite3())(expandedPath);
     }
   }
 
@@ -880,7 +881,7 @@ export function repairDatabase(): { status: 'ok' | 'repaired' | 'recreated'; mes
 
   // Full recreation
   const backupPath = backupCorruptDatabase(currentDbPath);
-  db = new BetterSqlite3(currentDbPath);
+  db = new (getBetterSqlite3())(currentDbPath);
   db.pragma('journal_mode = WAL');
   db.pragma('synchronous = NORMAL');
   db.pragma('foreign_keys = ON');
